refactor(basket): Get basket by userId, Store basket

+++ Services/Basket/YGZ.Basket.Application/Core/Abstractions/Payments/PaymentInformationModel.cs
@@ -5,7 +5,7 @@ namespace YGZ.Basket.Infrastructure.Payments.Vnpay;
 public class PaymentInformationModel
 {
     public string OrderType { get; set; } = default!;
-    public double Amount { get; set; } = default!;
+    public decimal Amount { get; set; } = default!;
     public string OrderDescription { get; set; } = default!;
     public string Name { get; set; } = default!;
 }
